+++ src/Controller/Game/GameGenreController.php
@@ -17,15 +17,18 @@ class GameGenreController extends AbstractController
     #[Route('/', name: 'game_genre', methods: ['GET'])]
     public function index(GameGenreRepository $gameGenreRepository): Response
     {
+
         return $this->render('game/game_genre/index.html.twig', [
             'game_genres' => $gameGenreRepository->findAll(),
             'navLinkId' => 'game_genre',
         ]);
     }
 
+
     #[Route('/new', name: 'game_genre_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+
         $gameGenre = new GameGenre();
         $form = $this->createForm(GameGenreType::class, $gameGenre);
         $form->handleRequest($request);
@@ -44,9 +47,11 @@ class GameGenreController extends AbstractController
         ]);
     }
 
+
     #[Route('/{id}/details', name: 'game_genre_show', methods: ['GET'])]
     public function show(GameGenreRepository $gameGenreRepository, int $id): Response
     {
+
         $gameGenre = $gameGenreRepository->find($id);
 
         return $this->render('game/game_genre/show.html.twig', [
@@ -55,9 +60,11 @@ class GameGenreController extends AbstractController
         ]);
     }
 
+
     #[Route('/{id}/edit', name: 'game_genre_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, EntityManagerInterface $entityManager, GameGenreRepository $gameGenreRepository, int $id): Response
     {
+
         $gameGenre = $gameGenreRepository->find($id);
 
         $form = $this->createForm(GameGenreType::class, $gameGenre);
@@ -76,17 +83,21 @@ class GameGenreController extends AbstractController
         ]);
     }
 
+
     #[Route('/{id}/delete', name: 'game_genre_delete', methods: ['POST'])]
     public function delete(Request $request, EntityManagerInterface $entityManager, GameGenreRepository $gameGenreRepository, int $id): Response
     {
+
         $gameGenre = $gameGenreRepository->find($id);
 
-        /** @var ?string $token */
-        $token = $request->request->get('_token');
+        if ($gameGenre) {
+            /** @var ?string $token */
+            $token = $request->request->get('_token');
 
-        if ($this->isCsrfTokenValid('delete'.$gameGenre->getId(), $token)) {
-            $entityManager->remove($gameGenre);
-            $entityManager->flush();
+            if ($this->isCsrfTokenValid('delete'.$gameGenre->getId(), $token)) {
+                $entityManager->remove($gameGenre);
+                $entityManager->flush();
+            }
         }
 
         return $this->redirectToRoute('game_genre', [], Response::HTTP_SEE_OTHER);
